Prepare upgrading to  rollup 3
Need to move rollup.config.js to rollup.config.mjs to explicitely
declarate it as an ES module. Also needed to remove the svelte.config.js
file and include its content inside rollup.config.mjs as it could not be
properly imported otherwise. This file was not really used standalone
anyway.

Updated the svelte-preprocess packages to make sure they can be imported
as ES modules.